+++ spec/spec_helper.rb
@@ -1,5 +1,6 @@
 require 'puppetlabs_spec_helper/module_spec_helper'
 require 'rspec-puppet-facts'
+require 'lib/spec_helpers'
 
 include RspecPuppetFacts
 
@@ -7,6 +8,6 @@ PUPPET_VERSION = Gem.loaded_specs['puppet'].version.to_s
 
 dir = File.expand_path(File.dirname(__FILE__))
 
-Dir["#{dir}/support/*.rb"].sort.each { |f| require f }
+Dir["#{dir}/shared_examples/*.rb"].sort.each { |f| require f }
 
 at_exit { RSpec::Puppet::Coverage.report! }